+++ src/stylesheets/AssesmentPage.css
@@ -1,0 +1,73 @@
+/* Container for the cards */
+.assessment-container {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 16px; /* Space between cards */
+    padding: 16px;
+    justify-content: center; /* Center align the cards */
+  }
+  
+  /* Individual card styling */
+  .card {
+    background-color: #f8f9fa;
+    border: 1px solid #e0e0e0;
+    border-radius: 8px;
+    width: 300px;
+    padding: 16px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+    transition: transform 0.3s, box-shadow 0.3s;
+  }
+  
+  .card:hover {
+    transform: translateY(-5px);
+    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.15);
+  }
+  
+  /* Title styling */
+  .card-title {
+    font-size: 1.5em;
+    margin-bottom: 8px;
+    color: #333;
+  }
+  
+  /* Description styling */
+  .card-desc {
+    font-size: 1em;
+    margin-bottom: 12px;
+    color: #666;
+  }
+  
+  /* Total questions styling */
+  .card-questions {
+    font-size: 0.9em;
+    margin-bottom: 16px;
+    color: #888;
+  }
+  
+  /* Start button styling */
+  .start-button {
+    padding: 8px 16px;
+    font-size: 1em;
+    color: white;
+    background-color: #007bff;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.3s;
+  }
+  
+  .start-button:hover {
+    background-color: #0056b3;
+  }
+  
+  /* Responsive design */
+  @media (max-width: 768px) {
+    .card {
+      width: 100%; /* Make cards full width on smaller screens */
+    }
+  }
+
